setup Flutter Clean Architecture with Provider
