+++ lib/data_layer/models/userModel.dart
@@ -7,17 +7,17 @@ class UserModel {
   final String userName;
   final String email;
 
-  UserModel(
-      {required this.uid,
-        required this.userName,
-        required this.email,
-        });
+  UserModel({
+    required this.uid,
+    required this.userName,
+    required this.email,
+  });
 
   Map<String, dynamic> toJson() => {
-    'uid': uid,
-    'userName': userName,
-    'email': email,
-  };
+        'uid': uid,
+        'userName': userName,
+        'email': email,
+      };
 
   static UserModel fromSnap(DocumentSnapshot snap) {
     var snapshot = snap.data() as Map<String, dynamic>;
@@ -27,5 +27,4 @@ class UserModel {
       email: snapshot["email"],
     );
   }
-
 }
