--- lib/example/variable.dart
@@ -1,4 +1,0 @@
-
-
-bool showTextEditingWidget = false;
-bool showTextEditingContainerFlag = false;
