+++ lib/main.dart
@@ -1,44 +1,22 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:grozziieapk/providers/text_editing_provider.dart';
+import 'package:grozziieapk/presentation_layer/providers/text_editing_provider.dart';
+import 'package:grozziieapk/presentation_layer/ui/created_label/created_label.dart';
+import 'package:grozziieapk/presentation_layer/ui/login_section/splash_screen.dart';
 import 'package:provider/provider.dart';
-import 'example/example.dart';
-import 'resourcs/firebase_options.dart';
+import 'domain_layer/resourcs/firebase_options.dart';
 
-void main() async{
+void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
-  runApp( const MyApp());
+  runApp(const MyApp());
 }
 
-// class MyApp extends StatelessWidget {
-//   const MyApp({super.key});
-//   @override
-//   Widget build(BuildContext context) {
-//     return ScreenUtilInit(
-//       designSize: const Size(360, 640),
-//       minTextAdapt: true,
-//       splitScreenMode: true,
-//       builder: (BuildContext context, Widget? child) {
-//         return MaterialApp(
-//           title: 'Grozziie',
-//           debugShowCheckedModeBanner: false,
-//           theme: ThemeData(
-//             primarySwatch: Colors.blue,
-//           ),
-//           home: child,
-//         );
-//       },
-//       child: CreateLabel(),
-//     );
-//   }
-// }
-
 class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -47,11 +25,8 @@ class MyApp extends StatelessWidget {
       minTextAdapt: true,
       splitScreenMode: true,
       builder: (BuildContext context, Widget? child) {
-        return MultiProvider(
-          providers: [
-            ChangeNotifierProvider(create: (_) => TextEditingProvider()),
-            // Add more providers here if needed
-          ],
+        return ChangeNotifierProvider<TextEditingProvider>(
+          create: (_) => TextEditingProvider(),
           child: MaterialApp(
             title: 'Grozziie',
             debugShowCheckedModeBanner: false,
@@ -62,10 +37,7 @@ class MyApp extends StatelessWidget {
           ),
         );
       },
-      child: CreateLabel(),
+      child: const CreateLabel(),
     );
   }
 }
-
-
-
