+++ lib/presentation_layer/ui/login_section/splash_screen.dart
@@ -2,9 +2,10 @@ import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:grozziieapk/ui/login_section/login_screen.dart';
 
-import '../../app_style.dart';
+import '../../../utils/app_style.dart';
+import 'login_screen.dart';
+
 
 
 class Splashscreen extends StatefulWidget {
