--- lib/ui/created_label/create_label.dart
@@ -1,36 +1,0 @@
-
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:grozziieapk/main.dart';
-import 'package:grozziieapk/ui/created_label/widgets/app_bar.dart';
-import 'package:grozziieapk/ui/created_label/widgets/bottom_app_bar.dart';
-import 'package:grozziieapk/ui/created_label/label_container/label_container.dart';
-import 'package:grozziieapk/ui/created_label/widgets/main_options_container.dart';
-
-/*class CreateLabel extends StatelessWidget {
-  const CreateLabel({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    double screenWidth = ScreenUtil().screenWidth;
-    return Scaffold(
-      backgroundColor: const Color(0xffFFFFFF),
-      appBar: MyAppBar().buildAppBar(context),
-      body: Container(
-        height: MediaQuery.of(context).size.height,
-        child: Column(
-          children: [
-            LabelContainer(screenWidth: screenWidth),
-            SizedBox(height: MediaQuery.of(context).size.height * 0.3),
-            Expanded(
-              child:
-                  OptionsContainer(context: context, screenWidth: screenWidth),
-            ),
-          ],
-        ),
-      ),
-      bottomNavigationBar: MyBottomAppBar().buildBottomAppBarButton(screenWidth),
-    );
-  }
-}*/
