+++ lib/utils/reuseable_class.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class ReuseAbleClass{
+  buildOptionRow(List<Widget> children) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceAround,
+      children: children,
+    );
+  }
+
+  buildIconButton(
+      String imagePath, String label, VoidCallback onPressed) {
+    return InkWell(
+      onTap: onPressed,
+      child: SizedBox(
+        width: 55.w,
+        child: Column(
+          children: [
+            IconButton(
+              onPressed: onPressed,
+              icon: Image.asset(
+                imagePath,
+                width: 24.w,
+                height: 24.h,
+              ),
+              color: Colors.grey,
+            ),
+            SizedBox(height: 5.h),
+            Text(
+              label,
+              textAlign: TextAlign.center,
+              style: TextStyle(
+                fontSize: 12.sp,
+                color: Colors.black45,
+                fontFamily: 'Poppins-Regular',
+                overflow: TextOverflow.ellipsis,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
