+++ pubspec.yaml
@@ -37,7 +37,6 @@ dependencies:
   cloud_firestore:
   firebase_auth:
   firebase_storage:
-  provider: 
 
 
   # The following adds the Cupertino Icons font to your application.
